requirements.txt
Django
django-ckeditor
django-crispy-forms
django-js-asset
django-ranged-response
django-simple-captcha
gunicorn
Pillow
psycopg2-binary
upsidedown